# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Class Pets API
  version: 1.0.0
  description: 班级宠物积分系统接口文档（Spring Boot + MyBatis-Plus + MySQL）。注意：本文档含规划接口与已实现接口，联调请以 PROJECT_STATUS.md 中“当前接口（真实）”为准。
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /classes/{classId}/students:
    get:
      tags: [Student]
      parameters:
        - { $ref: '#/components/parameters/ClassId' }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 20 }
        - in: query
          name: keyword
          schema: { type: string }
        - in: query
          name: groupId
          schema: { type: integer, format: int64 }
        - in: query
          name: sort
          schema: { type: string, enum: [points_desc, points_asc, name, update_desc] }
      responses:
        "200": { description: ok }
    post:
      tags: [Student]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentCreateRequest'
      responses:
        "200": { description: ok }
  /students/{studentId}:
    get:
      tags: [Student]
      parameters: [ { $ref: '#/components/parameters/StudentId' } ]
      responses:
        "200": { description: ok }
    put:
      tags: [Student]
      parameters: [ { $ref: '#/components/parameters/StudentId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentUpdateRequest'
      responses:
        "200": { description: ok }
    delete:
      tags: [Student]
      parameters: [ { $ref: '#/components/parameters/StudentId' } ]
      responses:
        "200": { description: ok }
  /students/{studentId}/score:
    post:
      tags: [StudentEvent]
      parameters: [ { $ref: '#/components/parameters/StudentId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentScoreRequest'
      responses:
        "200": { description: ok }
  /classes/{classId}/students/batch-score:
    post:
      tags: [StudentEvent]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchScoreRequest'
      responses:
        "200":
          description: ok
          content:
            application/json:
              example:
                code: 0
                message: "OK"
                data:
                  total: 3
                  successCount: 3
                  failedCount: 0
                  successIds: [1001, 1002, 1003]
...
  /classes/{classId}/groups:
    get:
      tags: [Group]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      responses:
        "200": { description: ok }
    post:
      tags: [Group]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupCreateRequest'
      responses:
        "200": { description: ok }
  /groups/{groupId}:
    put:
      tags: [Group]
      parameters: [ { $ref: '#/components/parameters/GroupId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupUpdateRequest'
      responses:
        "200": { description: ok }
    delete:
      tags: [Group]
      parameters: [ { $ref: '#/components/parameters/GroupId' } ]
      responses:
        "200": { description: ok }
  /groups/{groupId}/score:
    post:
      tags: [GroupEvent]
      parameters: [ { $ref: '#/components/parameters/GroupId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupScoreRequest'
      responses:
        "200": { description: ok }
  /classes/{classId}/rules:
    get:
      tags: [Rule]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      responses:
        "200": { description: ok }
    post:
      tags: [Rule]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreateRequest'
      responses:
        "200": { description: ok }
...
  /classes/{classId}/growth/config:
    get:
      tags: [Growth]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      responses:
        "200":
          description: ok
          content:
            application/json:
              example:
                code: 0
                message: "OK"
                data:
                  customized: true
                  overflowStep: 120
                  expGainRatio: 10
                  levelThresholds: [0, 100, 260, 480]
                  levelItems:
                    - level: 1
                      name: "萌芽"
                      threshold: 0
                    - level: 2
                      name: "成长"
                      threshold: 100
                  evolution:
                    stage1MaxLevel: 3
                    stage2MaxLevel: 6
                    stage3StartLevel: 7
                  petRoutes:
                    - id: "cat_route"
                      name: "喵咪路线"
                      enabled: true
                      stages:
                        - stage: 1
                          image: "/uploads/pets/cat-stage1.png"
                        - stage: 2
                          image: "/uploads/pets/cat-stage2.png"
    put:
      tags: [Growth]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrowthConfigUpdateRequest'
      responses:
        "200":
          description: ok
          content:
            application/json:
              example:
                code: 0
                message: "OK"
                data:
                  customized: true
                  overflowStep: 120
                  expGainRatio: 10
                  levelThresholds: [0, 100, 260, 480]
  /classes/{classId}/growth/recalculate:
    post:
      tags: [Growth]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      responses:
        "200":
          description: ok
          content:
            application/json:
              example:
                code: 0
                message: "OK"
                data: null
  /classes/{classId}/data/export/students:
    get:
      tags: [Data]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      responses:
        "200": { description: ok }
...
  /classes/{classId}/store/items:
    get:
      tags: [Store]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      responses:
        "200": { description: ok }
    post:
      tags: [Store]
      parameters: [ { $ref: '#/components/parameters/ClassId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreItemCreateRequest'
      responses:
        "200": { description: ok }
  /store/items/{itemId}:
    put:
      tags: [Store]
      parameters:
        - in: path
          name: itemId
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreItemUpdateRequest'
      responses:
        "200": { description: ok }
    delete:
      tags: [Store]
      parameters:
        - in: path
          name: itemId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200": { description: ok }
  /store/redeem:
    post:
      tags: [Store]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreRedeemRequest'
      responses:
        "200": { description: ok }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ClassId: { in: path, name: classId, required: true, schema: { type: integer, format: int64 } }
    StudentId: { in: path, name: studentId, required: true, schema: { type: integer, format: int64 } }
    GroupId: { in: path, name: groupId, required: true, schema: { type: integer, format: int64 } }
...
  schemas:
...
    GrowthConfigUpdateRequest:
      type: object
      properties:
        levelItems:
          type: array
          items:
            $ref: '#/components/schemas/GrowthLevelItem'
        levelThresholds:
          type: array
          items: { type: integer }
        overflowStep: { type: integer }
        expGainRatio: { type: number, format: double }
        evolution:
          $ref: '#/components/schemas/GrowthEvolutionConfig'
        petRoutes:
          type: array
          items:
            $ref: '#/components/schemas/GrowthPetRoute'
        recalculate: { type: boolean, default: true }
      example:
        levelItems:
          - level: 1
            name: "萌芽"
            threshold: 0
          - level: 2
            name: "成长"
            threshold: 100
          - level: 3
            name: "进阶"
            threshold: 260
        overflowStep: 120
        expGainRatio: 10
        evolution:
          stage1MaxLevel: 3
          stage2MaxLevel: 6
          stage3StartLevel: 7
        petRoutes:
          - id: "cat_route"
            name: "喵咪路线"
            enabled: true
            stages:
              - stage: 1
                image: "/uploads/pets/cat-stage1.png"
              - stage: 2
                image: "/uploads/pets/cat-stage2.png"
              - stage: 3
                image: "/uploads/pets/cat-stage3.png"
        recalculate: true
    GrowthLevelItem:
      type: object
      properties:
        level: { type: integer }
        name: { type: string }
        threshold: { type: integer }
    GrowthEvolutionConfig:
      type: object
      properties:
        stage1MaxLevel: { type: integer }
        stage2MaxLevel: { type: integer }
        stage3StartLevel: { type: integer }
    GrowthPetRoute:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        enabled: { type: boolean }
        stages:
          type: array
          items:
            $ref: '#/components/schemas/GrowthPetStage'
    GrowthPetStage:
      type: object
      properties:
        stage: { type: integer }
        image: { type: string }
    StudentCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        studentNo: { type: string }
        gender: { type: string, enum: [boy, girl, unknown] }
        groupId: { type: integer, format: int64 }
        avatarImage: { type: string }
        petId: { type: string }
    StudentUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/StudentCreateRequest'
        - type: object
          properties:
            totalPoints: { type: integer }
            redeemPoints: { type: integer }
            exp: { type: integer }
            level: { type: integer }
            title: { type: string }
    StudentScoreRequest:
      type: object
      required: [mode]
      properties:
        mode: { type: string, enum: [add, deduct] }
        ruleId: { type: integer, format: int64 }
        points: { type: integer }
        reason: { type: string }
        note: { type: string }
        context: { type: string, enum: [manual, roll], default: manual }
    BatchScoreRequest:
      type: object
      required: [studentIds, mode, points]
      properties:
        studentIds: { type: array, items: { type: integer, format: int64 } }
        mode: { type: string, enum: [add, deduct] }
        points: { type: integer }
        reason: { type: string }
      example:
        studentIds: [1001, 1002, 1003]
        mode: add
        points: 2
        reason: "课堂积极参与"
...
    GroupCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        icon: { type: string }
        note: { type: string }
        points: { type: integer }
    GroupUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/GroupCreateRequest'
    GroupScoreRequest:
      type: object
      required: [mode, points, reason]
      properties:
        mode: { type: string, enum: [add, deduct] }
        points: { type: integer }
        reason: { type: string }
        note: { type: string }
    RuleCreateRequest:
      type: object
      required: [content, points, type]
      properties:
        content: { type: string }
        points: { type: integer }
        type: { type: string, enum: [add, deduct] }
        category: { type: string }
        enabled: { type: boolean, default: true }
        cooldownHours: { type: number, format: float, default: 0 }
        stackable: { type: boolean, default: true }
    StoreItemCreateRequest:
      type: object
      required: [name, cost, stock]
      properties:
        name: { type: string }
        cost: { type: integer }
        stock: { type: integer }
        icon: { type: string }
        isActive: { type: boolean, default: true }
    StoreItemUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/StoreItemCreateRequest'
    StoreRedeemRequest:
      type: object
      required: [studentId, itemId]
      properties:
        studentId: { type: integer, format: int64 }
        itemId: { type: integer, format: int64 }